dns v1

+++ any_precision/quantization/main.py
@@ -33,8 +33,8 @@ def any_precision_quantize(
         random_state=None,
         group_count=1,
         dns=False,
-        sensitivity_outlier_percent=0.05,
-        threshold_outlier_percent=0.40
+        sensitivity_outlier_percent=0.00,
+        threshold_outlier_percent=0.45
 ):
     assert mode in ['gradients', 'quantize', 'pack'], \
         "mode must be one of 'gradients', 'quantize', or 'pack'. Use 'pack' to run the entire pipeline."
@@ -109,20 +109,22 @@ def any_precision_quantize(
     # ------------------- Dense & Sparse -------------------
 
     if dns:
-        logging.info("------------------- Dense & Sparse -------------------")
-        sparse_model_weights = remove_outliers(
-            analyzer=analyzer,
-            gradients=model_gradients,
-            sensitivity_outlier_percent=sensitivity_outlier_percent,
-            threshold_outlier_percent=threshold_outlier_percent,
-        )
-
         sparse_path = f"{quantized_cache_path}/sparse"
-        os.makedirs(sparse_path, exist_ok=True)
-        for l in range(analyzer.num_layers):
-            torch.save(sparse_model_weights[l], f"{sparse_path}/l{l}.pt")
+        if not os.path.exists(sparse_path) and not os.path.exists(f"{sparse_path}/l0.pt"):
+            logging.info("------------------- Dense & Sparse -------------------")
+            sparse_model_weights = remove_outliers(
+                analyzer=analyzer,
+                gradients=model_gradients,
+                sensitivity_outlier_percent=sensitivity_outlier_percent,
+                threshold_outlier_percent=threshold_outlier_percent,
+            )
+            os.makedirs(sparse_path, exist_ok=True)
+            for l in range(analyzer.num_layers):
+                torch.save(sparse_model_weights[l], f"{sparse_path}/l{l}.pt")
 
-        del sparse_model_weights
+            del sparse_model_weights
+        else:
+            logging.info(f"Sparse path {sparse_path} already exists and is not empty. Will skip dns.")
 
     # ------------------- Quantize: Seed + Upscale -------------------
